Propagation of exceptions added to RawVariantsQueryRunner.
