+++ dae/dae/query_variants/query_runners.py
@@ -77,6 +77,34 @@ class QueryRunner(abc.ABC):
     def _set_result_queue(self, result_queue):
         self._result_queue = result_queue
 
+    def _put_value_in_result_queue(self, val):
+        assert self._result_queue is not None
+
+        no_interest = 0
+        while True:
+            try:
+                self._result_queue.put(val, timeout=0.1)
+                break
+            except queue.Full:
+                logger.debug(
+                    "runner (%s) nobody interested",
+                    self.study_id)
+
+                if self.closed():
+                    break
+                no_interest += 1
+                if no_interest % 1_000 == 0:
+                    logger.warning(
+                        "runner (%s) nobody interested %s",
+                        self.study_id, no_interest)
+                if no_interest > 5_000:
+                    logger.warning(
+                        "runner (%s) nobody interested %s"
+                        "closing...",
+                        self.study_id, no_interest)
+                    self.close()
+                    break
+
 
 class QueryResult:
     """Run a list of queries in the background.
